rename sprite and imageobject to more idiomatic names
